DELIA-52496:Continuous aamp pos(current pos) prints seen for milliseconds with duplicate numbers
Reason for change: progressreportinterval configuration value from application need to be converted to milliseconds
Test Procedure: Play peacock content and check for aamp pos
Risks: Low

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_21_1
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_21_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_21_1

+++ aampgstplayer.cpp
@@ -617,9 +617,9 @@ static gboolean IdleCallback(gpointer user_data)
 		_this->privateContext->firstProgressCallbackIdleTaskId = 0;
 		if (0 == _this->privateContext->periodicProgressCallbackIdleTaskId)
 		{
-			 int  reportProgressInterval;
+			 double  reportProgressInterval;
 			 _this->aamp->mConfig->GetConfigValue(eAAMPConfig_ReportProgressInterval,reportProgressInterval);
-			_this->privateContext->periodicProgressCallbackIdleTaskId = g_timeout_add(reportProgressInterval, ProgressCallbackOnTimeout, user_data);
+			 _this->privateContext->periodicProgressCallbackIdleTaskId = g_timeout_add((int)reportProgressInterval, ProgressCallbackOnTimeout, user_data);
 			AAMPLOG_WARN("%s:%d current %d, periodicProgressCallbackIdleTaskId %d", __FUNCTION__, __LINE__, g_source_get_id(g_main_current_source()), _this->privateContext->periodicProgressCallbackIdleTaskId);
 		}
 		else
